% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Bar_plot_feature_importance}
\alias{F_Bar_plot_feature_importance}
\title{F_Bar_plot_feature_importance}
\usage{
F_Bar_plot_feature_importance(shp)
}
\arguments{
\item{shp}{= a shapviz object encapsulating the SHAP values}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a PNG bar plot of feature importance
}\if{html}{\out{</div>}}
}
\description{
This function generates a bar chart showing the importance of the 30
most important features of a machine learning model,
according to the calculated SHAP values.
}
\examples{
       plot <- F_Bar_plot_feature_importance(shp)
}
