% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Create_hypergrid}
\alias{F_Create_hypergrid}
\title{F_Create_hypergrid}
\usage{
F_Create_hypergrid()
}
\arguments{
\item{none}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   hypergrid
}\if{html}{\out{</div>}}
}
\description{
This function create the hypergrid values by user
}
\examples{
       hypergrid <- F_Create_hypergrid()
}
