% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Create_optimal_params}
\alias{F_Create_optimal_params}
\title{F_Create_optimal_params}
\usage{
F_Create_optimal_params()
}
\arguments{
\item{none}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   the optimal parameters
}\if{html}{\out{</div>}}
}
\description{
This function creates the optimal parameters with data
entered manually by the user.
}
\examples{
       optimal_params <- F_Create_optimal_params()
}
