% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Df_with_shap_values}
\alias{F_Df_with_shap_values}
\title{F_Df_with_shap_values}
\usage{
F_Df_with_shap_values(optimal_pars, dataframe)
}
\arguments{
\item{optimal_pars}{= the optimal parameter for the model
dataframe     = the standardize dataframe with the dependent variable (bias)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a dataframe of the original data in a long format
}\if{html}{\out{</div>}}
}
\description{
This function trains an xgboost model with optimized hyperparameters,
calculates SHAP values and then combines these values with the original data
in a long format that is useful for further analysis and visualization.
}
\examples{
       long_dataframe <- F_Df_with_shap_values(optimal_pars,dataframe)
}
