% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Extract_centroids}
\alias{F_Extract_centroids}
\title{F_Extract_centroids}
\usage{
F_Extract_centroids(dataframe)
}
\arguments{
\item{lda_sdf:A}{dataframe with geometry data, it must contain a column with the area code.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A dataframe with latitude and longitude variables representing the centroids.
}\if{html}{\out{</div>}}
}
\description{
This function extracts the centroids of a file with geometry of
municipalities or geographic areas.
}
\examples{
       data_centroids <- F_Extract_centroids(dataframe)
}
