% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Feature_importance_value_rankings}
\alias{F_Feature_importance_value_rankings}
\title{F_Feature_importance_value_rankings}
\usage{
F_Feature_importance_value_rankings(shp)
}
\arguments{
\item{shp}{= a shapviz object encapsulating the SHAP values}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a dataframe containing the feature names, their importance values and their ranking
}\if{html}{\out{</div>}}
}
\description{
This function generates a feature importance bar chart,
extracts the underlying data, calculates the ranking of the features
based on their importance values.
}
\examples{
       ranks <- F_Feature_importance_value_rankings(shp)
}
