% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Fit_final_model}
\alias{F_Fit_final_model}
\title{F_Fit_final_model}
\usage{
F_Fit_final_model(optimal_pars, X_train, y_train)
}
\arguments{
\item{optimal_pars:}{optimal parameters}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a trained model which is ready to make predictions
}\if{html}{\out{</div>}}
}
\description{
This function takes the optimal hyperparameters previously found,
trains a final xgboost model using those hyperparameters and returns
the trained model. The trained model is then ready to make predictions.
}
\examples{
       trained_model <- F_Fit_final_model(optimal_pars)
}
