% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Get_and_validate_numerical_input}
\alias{F_Get_and_validate_numerical_input}
\title{F_Get_and_validate_numerical_input}
\usage{
F_Get_and_validate_numerical_input(prompt, type, input_mode)
}
\arguments{
\item{prompt:}{message displayed to user
type  : a string "d" for decimals or "i" for integers
input_mode :}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   validated values to numeric format
}\if{html}{\out{</div>}}
}
\description{
Function to obtain and validate the input (integer or decimal) of the user
}
\examples{
       entry <- F_Get_and_validate_numerical_input("Enter a decimal value","d")
}
