% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Get_hypergrid}
\alias{F_Get_hypergrid}
\title{F_Get_hypergrid}
\usage{
F_Get_hypergrid(mode)
}
\arguments{
\item{mode}{: an integer value (1 or 2)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   it depends on mode:
   mode = 1 : returns values for test (32 combinations)
   mode = 2 : returns values for complete experiment (5184 combinations)
}\if{html}{\out{</div>}}
}
\description{
This function extracts the hypergrid values by default
}
\examples{
       hypergrid <- F_Get_hypergrid(1)
}
