% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Get_numerical_values_user}
\alias{F_Get_numerical_values_user}
\title{F_Get_numerical_values_user}
\usage{
F_Get_numerical_values_user(prompt)
}
\arguments{
\item{prompt}{: Text requesting user input}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a numeric value
}\if{html}{\out{</div>}}
}
\description{
Function to request a value from the user and validate that it is numeric
}
\examples{
       alpha_value <- F_Get_numerical_values_user(prompt)
}
