% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Get_selected_areas}
\alias{F_Get_selected_areas}
\title{F_Get_selected_areas}
\usage{
F_Get_selected_areas(dataframe)
}
\arguments{
\item{dataframe}{= the standardize dataframe with the dependent variable (bias)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a vector containing the names of the 20 areas or municipalities
}\if{html}{\out{</div>}}
}
\description{
This function identifies the 10 areas or municipalities with the highest and
lowest population, combines them, sorts them by descending population
and returns a vector containing the names of those municipalities.
}
\examples{
       areas <- F_Get_selected_areas(dataframe)
}
