% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Get_train_proportion}
\alias{F_Get_train_proportion}
\title{F_Get_train_proportion}
\usage{
F_Get_train_proportion()
}
\arguments{
\item{none}{}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   training sample proportion
}\if{html}{\out{</div>}}
}
\description{
This function get the train proportion by the user
}
\examples{
       train_sample_prop <- F_Get_train_proportion()
}
