% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Heatmap_sorted_alphabetically}
\alias{F_Heatmap_sorted_alphabetically}
\title{F_Heatmap_sorted_alphabetically}
\usage{
F_Heatmap_sorted_alphabetically(input_df_with_shap_values, selected_areas)
}
\arguments{
\item{input_df_with_shap_values}{= a dataframe of the original data in a long format
(returned by the F_Df_with_shap_values function)
selected_areas = a vector of selected municipalities or areas}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a PNG heat map showing the SHAP values sorted alphabetically for the selected municipalities
}\if{html}{\out{</div>}}
}
\description{
This function generates a heat map showing the SHAP values for the selected
municipalities and the corresponding characteristics sorted alphabetically.
}
\examples{
       plot <- F_Heatmap_sorted_alphabetically(input_df_with_shap_values,selected_areas)
}
