% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Is_integer}
\alias{F_Is_integer}
\title{F_Is_integer}
\usage{
F_Is_integer(string)
}
\arguments{
\item{a}{string}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   TRUE: If the complete string matches the regular expression
   FALSE: If the string does not match the regular expression
}\if{html}{\out{</div>}}
}
\description{
Function to validate if a string represents an integer number
}
\examples{
       number <- F_Is_decimal(string)
}
