% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Metrics_table}
\alias{F_Metrics_table}
\title{F_Metrics_table}
\usage{
F_Metrics_table(train_metrics, test_metrics)
}
\arguments{
\item{train_sample_metrics}{: metrics obtained from the predictions of the F_Predictions function
test_sample_metrics    : same as train_sample_metrics}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a summary table with metrics and correlations
}\if{html}{\out{</div>}}
}
\description{
This function creates the results table of performance metrics and correlations
}
\examples{
       summary_table <- F_Metrics_table(train_sample_metrics,test_sample_metrics)
}
