% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Missing_values}
\alias{F_Missing_values}
\title{F_Missing_values}
\usage{
F_Missing_values(dataframe)
}
\arguments{
\item{dataframe:A}{dataframe with or without missing values}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A dataframe without missing values
}\if{html}{\out{</div>}}
}
\description{
This function implements the listwise deletion approach as a
treatment for missing values and returns a dataframe without missing values.
}
\examples{
       complete_data <- F_Missing_values(dataframe)
}
