% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Parallelised_code}
\alias{F_Parallelised_code}
\title{F_Parallelised_code}
\usage{
F_Parallelised_code(hyper_grid, X_train, y_train)
}
\arguments{
\item{hyper_grid}{:data with the different combinations of parameters of the XGBoost model
features_train :independent variables
response_train :dependent variable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a dataframe with the results for each hyperparameter
}\if{html}{\out{</div>}}
}
\description{
This function performs a parallel grid search to find the best combination
of hyperparameters for an xgboost regression model, using cross-validation
to evaluate performance. The code measures run time and returns a dataframe
with the results for each hyperparameter combination.
}
\examples{
       results <- F_Parallelised_code(hyper_grid,features_train,response_train)
}
