% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Partial_dependence_plots}
\alias{F_Partial_dependence_plots}
\title{F_Partial_dependence_plots}
\usage{
F_Partial_dependence_plots(shp, xgb.fit.final, features_train, shap_values)
}
\arguments{
\item{shp}{= a shapviz object encapsulating the SHAP values
xgb.fit.final   = the trained model
features_train  = the independent variables of the training set
shap_values     = The shap values}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a PNG grid of scatter plots
}\if{html}{\out{</div>}}
}
\description{
This function generates a grid of scatter plots showing the relationship
between the values of the 8 most important features and their SHAP values,
providing information on how each feature influences the model predictions.
}
\examples{
       plot <- F_Partial_dependence_plots(shp,xgb.fit.final,features_train,shap_values)
}
