% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Predictions}
\alias{F_Predictions}
\title{F_Predictions}
\usage{
F_Predictions(xgb.fit.final, sample, bias_or_metrics)
}
\arguments{
\item{xgb.fit.final}{: trained model
sample          : training set or test set
bias_or_metrics : 1 returns sample with predictions / 2 returns prediction's metrics}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   depends on the parameter bias_or_metrics:
   bias_or_metrics = 1 : returns the data set with an extra column named "bias_prediction" with the predictions
   bias_or_metrics = 2 : returns a vector with the columns: rmse,pearson correlation,spearmam correlation,standard deviation
}\if{html}{\out{</div>}}
}
\description{
This function uses the model, the data set and makes predictions.
}
\examples{
       predictions <- F_Predictions(model,test_data,1)
}
