% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Shap_plot_bias}
\alias{F_Shap_plot_bias}
\title{F_Shap_plot_bias}
\usage{
F_Shap_plot_bias(shp)
}
\arguments{
\item{shp}{= a shapviz object encapsulating the SHAP values}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a PNG swarm plot of SHAP values
}\if{html}{\out{</div>}}
}
\description{
This function generates a swarm plot showing the distribution of SHAP values
for the 20 most important features of a model.
}
\examples{
       plot <- F_Shap_plot_bias(shp)
}
