% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Standardize}
\alias{F_Standardize}
\title{F_Standardize}
\usage{
F_Standardize(dataframe)
}
\arguments{
\item{dataframe:A}{dataframe with mixed attributes (numerical and categorical)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A dataframe with standardized numeric attributes
}\if{html}{\out{</div>}}
}
\description{
This function standardizes the numeric attributes of a dataframe.
}
\examples{
       standardized_data <- F_Standardize(dataframe)
}
