% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Training}
\alias{F_Training}
\title{F_Training}
\usage{
F_Training(X_train, y_train)
}
\arguments{
\item{features_train}{:independent variables
response_train :dependent variable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Plot of model performance as the number of trees increases
}\if{html}{\out{</div>}}
}
\description{
This function trains a Gradient Boosting model using cross-validation,
finds the optimal number of trees, and visualizes the performance of the model
as the number of trees increases. This allows the model to be evaluated and optimized.
}
\examples{
       F_Training(features_train,response_train)
}
