% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Explain_Bias.R
\name{F_Treatment_plan}
\alias{F_Treatment_plan}
\title{F_Treatment_plan}
\usage{
F_Treatment_plan(features, train_sample, test_sample, partition)
}
\arguments{
\item{features}{:list of attributes or independent variables
train_sample :dataframe training set
test_sample  :dataframe test set
partition    :Processed partition to be returned (1 = training set, 2 = test set)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A Dataframe with optimized attributes or variables
}\if{html}{\out{</div>}}
}
\description{
This function implements a treatment plan for the data. It performs
transformations to both numerical and categorical variables to optimize
their use in the model.
}
\examples{
       optimized_data <- F_Treatment_plan(independent_variables,training_set,test_set,1)
}
